backend/code/requirements.txt
Django==1.11.5
psycopg2==2.7.3.1
py==1.4.34
pytest==3.2.2
pytest-django==3.1.2
pytz==2017.2
facebook-sdk